requirements.txt
google-adk==1.28.0
langchain-community
langchain-google-vertexai
google-cloud-discoveryengine>=0.13.12,<0.14.0
wikipedia
beautifulsoup4
requests>=2.32.4
google-cloud-translate
google-cloud-speech
python-dotenv
pandas